// plugins/plugin-harbor/src/utils/interfaces.ts
export interface IArtifact {
  digest: string;
  // eslint-disable-next-line @typescript-eslint/naming-convention
  extra_attrs: {
    architecture: string;
    author: string;
    config: {
      Cmd?: string[];
      Entrypoint?: string[];
      Env?: string[];
      // eslint-disable-next-line @typescript-eslint/no-explicit-any
      ExposedPorts?: { [key: string]: any };
      Labels?: { [key: string]: string };
      User?: string;
      WorkingDir?: string;
    };
    created: string;
    os: string;
  };
  icon: string;
  id: number;
  // eslint-disable-next-line @typescript-eslint/naming-convention
  manifest_media_type: string;
  // eslint-disable-next-line @typescript-eslint/naming-convention
  media_type: string;
  // eslint-disable-next-line @typescript-eslint/naming-convention
  project_id: number;
  // eslint-disable-next-line @typescript-eslint/naming-convention
  pull_time: string;
  // eslint-disable-next-line @typescript-eslint/naming-convention
  push_time: string;
  // eslint-disable-next-line @typescript-eslint/no-explicit-any
  references?: IArtifactReference[];
  // eslint-disable-next-line @typescript-eslint/naming-convention
  repository_id: number;
  // eslint-disable-next-line @typescript-eslint/naming-convention
  scan_overview: { [key: string]: IArtifactScanOverview };
  size: number;
  tags: IArtifactTag[];
  type: string;
}

export interface IArtifactReference {
  // eslint-disable-next-line @typescript-eslint/naming-convention
  child_digest: string;
  // eslint-disable-next-line @typescript-eslint/naming-convention
  child_id: number;
  // eslint-disable-next-line @typescript-eslint/naming-convention
  parent_id: number;
  platform: {
    OsFeatures: string[];
    architecture: string;
    os: string;
  };
  urls: string[];
}

export interface IArtifactScanOverview {
  // eslint-disable-next-line @typescript-eslint/naming-convention
  complete_percent: number;
  duration: number;
  // eslint-disable-next-line @typescript-eslint/naming-convention
  end_time: string;
  // eslint-disable-next-line @typescript-eslint/naming-convention
  report_id: string;
  // eslint-disable-next-line @typescript-eslint/naming-convention
  scan_status: string;
  severity: string;
  // eslint-disable-next-line @typescript-eslint/naming-convention
  start_time: string;
}

export interface IArtifactTag {
  // eslint-disable-next-line @typescript-eslint/naming-convention
  artifact_id: number;
  id: number;
  immutable: boolean;
  name: string;
  // eslint-disable-next-line @typescript-eslint/naming-convention
  pull_time: string;
  // eslint-disable-next-line @typescript-eslint/naming-convention
  push_time: string;
  // eslint-disable-next-line @typescript-eslint/naming-convention
  repository_id: number;
  signed: boolean;
}